requirements.txt
pynput==1.7.6
pandas==2.0.3
Pillow==10.0.0
openpyxl==3.1.2